Rename module pydantic_models
